Move helper functions into _vectorops
